# perf(serialize): faster object entries serialization