--- a/src/serialize.ts
+++ b/src/serialize.ts
@@ -116,15 +116,13 @@ const Serializer = /*@__PURE__*/ (function () {
         return handler.call(this, object);
       }
       if (typeof object?.entries === "function") {
-        return this.serializeObjectEntries(type, object.entries());
+        return this.serializeObjectEntries(type, Array.from(object.entries()));
       }
       throw new Error(`Cannot serialize ${type}`);
     }
 
-    serializeObjectEntries(type: string, entries: Iterable<[any, any]>) {
-      const sortedEntries = Array.from(entries).sort((a, b) =>
-        this.compare(a[0], b[0]),
-      );
+    serializeObjectEntries(type: string, entries: Array<[any, any]>) {
+      const sortedEntries = entries.sort((a, b) => this.compare(a[0], b[0]));
       let content = `${type}{`;
       for (let i = 0; i < sortedEntries.length; i++) {
         const [key, value] = sortedEntries[i];
@@ -186,7 +184,7 @@ const Serializer = /*@__PURE__*/ (function () {
     }
 
     $Map(map: Map<any, any>) {
-      return this.serializeObjectEntries("Map", map.entries());
+      return this.serializeObjectEntries("Map", Array.from(map.entries()));
     }
   }